src/requirements-test.txt
mypy
nose2
nose2[coverage_plugin]>=0.6.5
black
factory-boy==2.9.2
